Workaround workload `rvv-bench`.
This workload reads rdcycles to terminate the program if it runs for a
time that is too long. This behavior is causing illegal instructions on
NEMU. It also have an option of using a system call to obtain the number
of cycles, but this will cause the program to stall. This commit
workaround this problem by writing a dummy `rdcycles` function that just
returns zeros.