Changed image layout for `libcheckpoint`.
`fw_jump` instead of `fw_payload` is used for OpenSBI. The device tree
is place at 512KiB offset, and OpenSBI is placed at 1 MiB offset to make
space for `libcheckpoint` or other customs code. A minimal startup code
is added to start OpenSBI and pass device tree address to it.