Added download file caching and sha256 checking.
In this commit, the format of `links.txt` of each file is changed to
`file_name link sha256sum` for each line. A new script is written to
download and verify the files. When all files are downloaded and
verified, a sentinel file is touched to indicate these files are ready.