Added utility `nemu-exec`.
This commit added an ultiliy program to the built-in initramfs of the
Linux kernel. It disables NEMU timer interrupt, makes NEMU to enter
simpoint profiling mode, starts the workload, and stops NEMU with the
exit code of the workload.

This commit also rnames the original `nemu-halt` to `nemu-trap`. The new
`nemu-trap` now does not enter a forever loop any more if NEMU does not
stop after the debug call. This enabbles it to be used to disable NEMU
timer interrupt or make NEMU to enter the simpoint mode.