Worked around build problems of coremark workloads.
`coremark` and `coremark-pro` requires `/bin/sh` to be bash. This can
cause problems on Debian based distros, where `/bin/sh` is by defualt
dash. This project used `update-alternatives` in CI to workaround it,
but this cannot solve the problem when you build this project on your
own machine. This commit removed `update-alternatives` in CI, and
specifies `SHELL=bash` in the build scripts when involking `make`.